app/keiba_ai/pyproject.toml
[project]
name = "keiba-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.54.0,<2.0.0)",
    "google-cloud-run (>=0.15.0,<0.16.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "playwright (>=1.58.0,<2.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)"
]

[tool.poetry]
packages = [{include = "keiba_ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
